[성진] dto, repository 수정

+++ src/main/java/com/kh/billida/rentalHistory/model/dto/Rental.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Rental {
 	
 	private long historyIndex;
-	private int lockerId;
+	private Long lockerId;
 	private String userCode;
 	private Date rentStart;
 	private Date rentEnd;
